Judgment that the business is initially completed

+++ judge/models/judge_info.go
@@ -2,6 +2,6 @@ package models
 
 type JudgeInfo struct {
 	Message string
-	Memory  int64
-	Time    int64
+	Memory  int
+	Time    int
 }
